yarn-plugin: import appropriate type to verify beforeWorkspacePacking signature
I realised that by default the `Plugin` type from
`@yarnpkg/core` sets the type of the `hooks`
property to `any`. This meant that there wasn't
any restriction on the hooks that could be passed,
and introduced the risk that a typo in the
`beforeWorkspacePacking` name could lead to the
hook being skipped.

This commit pulls in the appropriate type for
Hooks to verify the name and signature of the hook.

Signed-off-by: MT Lewis <mtlewis@users.noreply.github.com>