Make GitHub environment Scaffolder action use auth to resolve reviewers
Changes the github:environment:create Scaffolder action to request and
use a backend auth token when resolving the reviewer entityRefs from the
Backstage catalog.

This is because previously it would throw a 401 error when backend auth
was not disabled.

The logic for requesting the token is copied from the existing
catalog:fetch action, which needs to do a similar thing.

Also slightly clarifies that Backstage entityRefs are expected in the
reviewers list for this action.

Signed-off-by: Jason Liu <a.jason.liu21@gmail.com>