Allow for auth provider logout
There are some proxied auth providers that require additional logout
steps. This change allows for all auth providers to provide an optional
`logout` method that may be called upon signing out from the frontend.

The logout method should respond with a LogoutResponse that may provide
a redirectUrl, which if set, directs the frontend to set
window.location.href.

Signed-off-by: Craig Tracey <craig@arctir.com>