canon: use inset box-shadows for button styles
The use of box-shadows for button borders meant
that when a button was styled with a border, it
expanded by 1px. This commit retains the use of
shadows for borders, but switches them to inset
shadows so that the component size stays
consistent between buttons with and without
borders (and between the same button in hover and
non-hover states).

Signed-off-by: MT Lewis <mtlewis@users.noreply.github.com>