Update provider.md
Corrected the Callback URL since we encountered this when implementing.  Also removed the scopes config item - including this resulted in errors for us.

I had submitted a PR before, and I addressed two feedback items - one was wording and the other was a request to include the optional configurations in the instructions.  Both are addressed.  Apologize I had to make a new PR because I had issues with the other being open so long and I didn't rebase it correctly - better to start fresh.

Signed-off-by: jolies93 <64967243+jolies93@users.noreply.github.com>