feat(techdocs): generator timeout config
this allows specifying techdocs generator timeout in configuration that
prevents techdocs generation to take infinite time in some cases like
with huge `plantuml` diagrams and local `mkdocs` renderer.

it does not fix the potential memory leak #27347 but might improve the
situation so that sockets get closed in case the generating is blocking
the system.

the timeout works both with the default docker generator as well as with
the local `mkdocs`.

Signed-off-by: Hellgren Heikki <heikki.hellgren@op.fi>