Allow disabling the header divider on homepage cards
Some homepage cards would like to define their own header
for example, so that a title can be set dynamically.

Removing the divider with the divider prop provides a way to put
everything in the content area without an extraneous divider above.

The divider is usually not easily visible using the default theme,
however customized themes that may use a larger border-radius this
divider becomes visible and distracting.

Signed-off-by: Adam Kunicki <kunickiaj@gmail.com>