Canon - Pin Base UI version (#29782)
Canon is built on Base UI primitives, but Base UI is still in alpha and does not follow semantic versioning. This means that breaking changes can appear in non-major dependency versions. As Canon does not currently depend on a specific version of Base UI, any project that consumes Canon that also installs a later version of Base UI could end up using a version of Base UI that Canon does not support, resulting in unexpected behaviour or bugs. To try and avoid this scenario, this PR pins the version of Base UI to the one that Canon has been developed against, meaning that Canon will install that version of Base UI rather than inheriting another if it is also in use elsewhere, and ensuring that the components all work as expected

---------

Signed-off-by: James Brooks <jamesbrooks@spotify.com>