Canon - Styling fixes for field clear button (#29878)
I spotted a couple of problems with the new clear button added to the TextField in #29820. Firstly, it did not use the correct colour token and therefore was not visible in dark mode. Secondly, for fields without a fixed width, the field would grow/shrink in size as the button was shown or hidden. Both of these issues have been rectified here

Signed-off-by: Rudra-SH <rudra099999@gmail.com>