Fix refresh API call, reorder refresh and delete buttons
`/api/catalog/refresh` responds with `200 OK` but does not send JSON data, so we rework the `CatalogUnprocessedEneitiesClient.fetch()` function to only parse the JSON from a response if the response's `Content-Type` header explicitly specifies that it contains JSON data.

All the existing API calls done by `CatalogUnprocessedEneitiesClient` behave the same with the above changes.

Reordered the refresh and delete buttons. Refresh is first, delete is second.

Signed-off-by: Kevin Snyder <kevinsnyder@Kevins-MacBook-Pro.local>
Signed-off-by: Kevin Snyder <kevin.snyder@gusto.com>