Match refresh logic of `AboutCart` with some differences
https://github.com/backstage/backstage/blob/176d5eb26a7df4ff37a36c47cbe0682c65bad1c2/plugins/catalog/src/components/AboutCard/AboutCard.tsx#L192-L203

Unlike the `AboutCard`, we do not check for permissions using the `useEntityPermission()` hook. This requires the current entity to be captured in context, which does not happen with the "Unprocessed Entities" page. Additionally, since unprocessed entities are hidden behind the "DevTools", and since we are not checking for permission for the delete action either, it seems acceptable to not do any additional permissions checks.

Another difference is that the `AboutCard` limits refreshes to locations that begin with `url:` or `file:` as they are less expensive. This makes sense for a view that is made accessible to a wider group of users, but for the "DevTools" page where users with elevated permissions will be using the refresh to debug and fix issues, we should allow refreshes for all locations, regardless of how expensive it is.

Signed-off-by: Kevin Snyder <kevin.snyder@gusto.com>