Reduce API calls made for single user events in GitHub Org Provider
Currently the provider will load all users group and all members of the groups when responding to an event on a single user. This can cause excessive API calls to page through group members when this information is just discarded in this instance.

This change introduces single user versions of the code that will not page through all membership.

Signed-off-by: Scott Guymer <scott.guymer@philips.com>