fix(catalog): processor order not to affect hash generation
sort the arrays such as relations, deferred entities, refresh keys and
parents coming from processors before updating the entity hash. when the
order of the processors is random, the order of the emitted relations
can be also random making unnecessary stitching occur for the entity as
the hash has changed.

this change most probably means that existing hashes for entities do not
match on the first processing round but makes the upcoming processing
rounds much more stable thus making the catalog processing more
performant.

Signed-off-by: Hellgren Heikki <heikki.hellgren@op.fi>