fix(catalog): msgraph to use only first parent for groups
the order of the groups from msgraph is random, thus the parent group
for groups having multiple parents can be random. this leads to
situation where the grpup parent is changing over the msgraph
provider runs leading to unnecessary stitching.

other possibility for the fix would be to change the model to allow
multiple parents for groups (like mentioned in the TODO) but that
could have some weird consequences in apps.

Signed-off-by: Hellgren Heikki <heikki.hellgren@op.fi>