feat(search): add actions to query search engine
two new actions registered to actions registry that allow fetching
supported search types and querying the search engine.

the authorized search engine initialization had to be moved to plugin
from router as it should also be used in the query action if permissions
are enabled.

additionally, a fix was needed for the mock actions service as the
invoke method return type only allowed JsonValue that doesn't match with
the IndexableResultSet with typescript error:
Index signature for type 'string' is missing in type 'ResultSet<IndexableDocument>'

also writing tests gets a bit easier with this as there is no need to
check for null values that JsonValue allowed.

Signed-off-by: Hellgren Heikki <heikki.hellgren@op.fi>