fix(scaffolder): distributed actions to be available
depending on the plugin startup order, some distributed actions might
not be available when the scaffolder backend starts. this was mentioned
in the comment that is now removed. while this might change in the
future, this would patch the current way the distributed actions are
fetched.

basically every time a task is started and when listing of the actions
is requested, the distributed actions will get a refresh. while it's
still possible that the distributed action suddenly disappears between
the start of the task and when the action is fetched from the
actionregistry, it's very slight chance of this happening.

closes #31187

Signed-off-by: Hellgren Heikki <heikki.hellgren@op.fi>