Replace package aws-sdk/signature-v4 with smithy
This package is not modified since 2 years now, and the
replacement is also mentioned in the documentation: https://www.npmjs.com/package/@aws-sdk/signature-v4?activeTab=readme

This PR is simply replacing the packages in the code, this is
a drop-in replacement and there's no need to apply any other changes.

This has been already done in one of the open-source plugins we own,
as well as in our Backstage instance, and all worked just like before.

Signed-off-by: ivangonzalezacuna <ivangonzalezacuna@gmail.com>