New technical overview per feature #21925 (#30998)
* Create technical-overview.md

First draft of technical overview. Just want to save these changes as I lost the last file when I had to reboot.

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Update technical-overview.md with additional text and graphics

Added additional text to describe the system model. Added graphics and started to add additional links.

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Update technical-overview.md added more text

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Update technical-overview.md add additional text to plugin section

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Update technical-overview.md add plugin architecture overview

Added text for plugin architecture overview and additional text in system model and purpose section. Added links for core features.

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Create backstage-ui-group-ownership.png

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Add files via upload

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Update technical-overview.md added graphics

Added a graphic illustrating My Group in the Backstage UI.

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Update technical-overview.md formatting changes

Made formatting changes. Changed bullets from + to - and some spacing changes trying to fix prettier errors.

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Update technical-overview.md copied in text from vscode running prettier

Trying to fix prettier errors. Edited code in vscode and save with prettier formatting. Replaced text with the vscode version.

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Update technical-overview.md update links for search and techdocs

There is a README.md file in each of these code directories (search and techdocs) but the online documentation displays the README file as index.html file for each of these features. So I replaced the links to the following:
https://backstage.io/docs/features/search/
https://backstage.io/docs/features/techdocs/

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Update technical-overview.md added graphic descriptions

Added "description" for graphics for accessibility.

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Update technical-overview.md update links

Added a link for "plugin architecture" to go to "plugin architecture" section in the architectural overview document.

Also moved the software catalog "entities" links around.

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Update sidebars.ts to add technical overview and remove some docs

Added the technical overview to the Overview section. Removed the what-is-backstage, background, and vision pages as that information was folded into the new technical overview.

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Update sidebars.ts to add back what-is-backstage and roadmap

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

* Update technical-overview.md updated text

Fixed spelling error and updated text to incorporate review comments.

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>

---------

Signed-off-by: milliehartnt123 <108491788+milliehartnt123@users.noreply.github.com>
Signed-off-by: Christine Lee <christine.lee@datadoghq.com>