Introduce new option in the GH catalog plugin to exclude suspended users
Introduce an optional setting to exclude suspended users from GitHub Enterprise instances.
When it’s set to true, suspended users won’t be emitted by the default transform.
If a custom transformer is used, it should check if the property `suspendedAt` from the `GithubUser` is defined,
in order to exclude such users. This logic was not introduced in the `GithubMultiOrgReaderProcessor.ts`,
since the usage there is marked as deprecated.

To note that this setting should be used only against GitHub Enterprise instances,
the property does not exist in the github.com GraphQL schema, adding it will cause a schema validation error
and the syncing of users will fail.

Signed-off-by: Valério Valério <vdv100@gmail.com>