feat: Expand `EntityFilter.getCatalogFilters` to return `EntityFilterQuery`
This change allows `EntityFilter`s to utilize more complete functionality of the `CatalogApi`, specifically the `queryEntities` and `getEntities` methods which support an `EntityFilterQuery` as the `filter` parameter in the request.

Allowing `EntityFilter`s to return a complete `EntityFilterQuery` allows them to produce more complex filter logic.

The primary example is allowing an `EntityFilter` to produce an array of filter records that would produce an 'OR' query, filtering on A OR B criteria.

To complete this change, some of the helpers and utilities used by Entity filtering functionality (ex: `useEntityListProvider`) needed to be updated to account for the possibility of an array of filter records.

Signed-off-by: Tim Klever <tim.v.klever@aexp.com>