[catalog-unprocessed-entities] Move types and clients from frontend to common plugin
Moved some types, as well as the API and client definitions to the common package.
The original types, client and interface are now re-exported from the common package instead.

This allows not only frontend plugins, but also backend ones, to use this client
if needed. Until now, and since the types and all definitions were stored in the
frontend plugin, the only way to implement such feature was to copy the code
in the project. It would make much more sense to make this exportable from the
plugin itself and let people use this plugin if needed.

I made this change a patch, since all types have been initially deprecated and
a mention to the new location has been added.

Signed-off-by: ivangonzalezacuna <ivangonzalezacuna@gmail.com>